+++ api-module/build.gradle
@@ -7,12 +7,6 @@ apply from: rootProject.file('gradle/dependencies.gradle')
 jar.enabled = true
 bootJar.enabled = false
 
-configurations {
-    compileOnly {
-        extendsFrom annotationProcessor
-    }
-}
-
 allure {
     version = "2.13.8"
     useJUnit5{
@@ -55,5 +49,4 @@ tasks.withType(Test) {
     testLogging {
         events "passed", "skipped", "failed"
     }
-    include "**/*Test.class"
 }
