+++ api-module/src/main/java/com/automation/api/configuration/TestConfig.java
@@ -5,14 +5,12 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 
-import static io.restassured.RestAssured.config;
-
 @Configuration
 @ComponentScan({"com.automation.api", "io.qameta.allure"})
 public class TestConfig {
 
     @Bean
     public RestAssuredConfig getRestAssuredConfig() {
-        return config();
+        return RestAssuredConfig.config();
     }
 }
