--- api-module/src/main/java/com/automation/api/models/enums/EntryType.java
@@ -1,5 +1,0 @@
-package com.automation.api.models.enums;
-
-public enum EntryType {
-    PERSONAL
-}
