+++ api-module/src/main/java/com/automation/api/models/userConstants/AccountType.java
@@ -1,0 +1,5 @@
+package com.automation.api.models.userConstants;
+
+public enum AccountType {
+    INTERNAL
+}
