+++ api-module/src/main/java/com/automation/api/models/userConstants/UserRole.java
@@ -1,0 +1,5 @@
+package com.automation.api.models.userConstants;
+
+public enum UserRole {
+    ADMINISTRATOR
+}
