+++ gradle/dependencies.gradle
@@ -1,8 +1,6 @@
 repositories {
     jcenter()
     mavenCentral()
-    maven { url "https://dl.bintray.com/epam/reportportal" }
-    maven { url "https://jitpack.io" }
 }
 
 ext {
